+++ narrator/text/text.py
@@ -1,20 +1,18 @@
 from __future__ import annotations
 import os
-from enum import Enum
-import string
 import datetime
 
-from narrator.sub_utils import blb2txt, balcon, ffmpeg__to_mp3
+from narrator.thirdparty import blb2txt, balcon, ffmpeg__to_mp3
+from narrator.utils import make_filename
 from narrator.exceptions import TextException
+from narrator.text.language import Language
+
 import narrator.config
 
 conf = narrator.config.text
 
 
 class Text:
-    class Language(Enum):
-        RU: str = "ru"
-        EN: str = "en"
 
     INPUT_FORMATS = conf.input_formats
 
@@ -33,36 +31,32 @@ class Text:
                     break
             text = "".join(to_join)
 
-        en_set = string.ascii_letters
-        ru_set = "абвгдеёжзийклмнопрстуфхцчшщъыьэюя"
-        ru_set += ru_set.upper()
-
         char_count = 0
-        score = {"en": 0, "ru": 0}
+        score = {lang.value: 0 for lang in Language}
         for ch in text:
-            if ch in en_set:
-                score["en"] += 1
-                char_count += 1
-            elif ch in ru_set:
-                score["ru"] += 1
-                char_count += 1
-            if char_count >= stop_at:
-                break
-        return Text.Language.EN if score["en"] >= score["ru"] else Text.Language.RU
+            lang_val = Language.get_val(ch)
+            if lang_val:
+                score[lang_val] += 1
+                if lang_val:
+                    char_count += 1
+                if char_count >= stop_at:
+                    break
+        _, max_lang_val = max((cnt, lang_val) for lang_val, cnt in score.items())
+        return Language(max_lang_val)
 
     def __init__(
-        self, title: str, paragraphs: list[str] | str, lang: Text.Language | None = None
+        self, title: str, paragraphs: list[str] | str, lang: Language | None = None
     ) -> None:
         self._title = title
         self._paragraphs = paragraphs
-        self._lang: Text.Language = lang or Text.guess_language(paragraphs)
+        self._lang: Language = lang or Text.guess_language(paragraphs)
         self._datetime = datetime.datetime.now()
 
     def _preamble(self) -> str:
         preamble = []
 
         preamble.append(self._title)
-        if self._lang == Text.Language.RU:
+        if self._lang == Language.RU:
             preamble.append("Создано: " + self._datetime.strftime("%Y-%m-%d %H:%M:%S"))
         else:
             preamble.append("Created: " + self._datetime.strftime("%Y-%m-%d %H:%M:%S"))
@@ -72,7 +66,7 @@ class Text:
         self, directory: str, filename: str | None = None, encoding: str = "utf-8"
     ) -> str:
         if not filename:
-            filename = self._title + ".txt"
+            filename = make_filename(self._title + ".txt")
         file_path = os.path.join(directory, filename)
         with open(file_path, mode="w", encoding=encoding) as txt:  # utf-8-sig
             txt.write(self._preamble() + "\n")
